Fix Everything
Rename to Fluid Extension API
Fix license and style
Add suggestions from review
