+++ fabric-fluid-extension-api-v1/build.gradle
@@ -1,0 +1,2 @@
+archivesBaseName = "fabric-fluid-extension-api-v1"
+version = getSubprojectVersion(project, "1.0.0")
