+++ fabric-fluid-extension-api-v1/src/main/resources/assets/fabric-fluid-extension-api-v1/icon.png
