+++ fabric-fluid-extension-api-v1/src/main/resources/fabric-fluid-extension-api-v1.mixins.json
