+++ fabric-fluid-extension-api-v1/src/testmod/java/net/fabricmc/fabric/test/fluid/extension/LemonadeFluid.java
@@ -1,0 +1,84 @@
+/*
+ * Copyright (c) 2016, 2017, 2018, 2019 FabricMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package net.fabricmc.fabric.test.fluid.extension;
+
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.entity.BlockEntity;
+import net.minecraft.fluid.Fluid;
+import net.minecraft.fluid.FluidState;
+import net.minecraft.item.Item;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.Direction;
+import net.minecraft.world.BlockView;
+import net.minecraft.world.WorldAccess;
+import net.minecraft.world.WorldView;
+
+import net.fabricmc.fabric.api.fluid.extension.v1.FabricFlowableFluid;
+
+public class LemonadeFluid extends FabricFlowableFluid {
+	@Override
+	public Block getFluidBlock() {
+		return FluidTest.LEMONADE_BLOCK;
+	}
+
+	@Override
+	public int getLevels() {
+		return 20;
+	}
+
+	@Override
+	protected boolean isInfinite() {
+		return false;
+	}
+
+	@Override
+	protected void beforeBreakingBlock(WorldAccess world, BlockPos pos, BlockState state) {
+		final BlockEntity blockEntity = state.getBlock().hasBlockEntity() ? world.getBlockEntity(pos) : null;
+		Block.dropStacks(state, world, pos, blockEntity);
+	}
+
+	@Override
+	protected int getFlowSpeed(WorldView world) {
+		return 4;
+	}
+
+	@Override
+	protected int getLevelDecreasePerBlock(WorldView world) {
+		return 1;
+	}
+
+	@Override
+	public Item getBucketItem() {
+		return FluidTest.LEMONADE_BUCKET;
+	}
+
+	@Override
+	protected boolean canBeReplacedWith(FluidState state, BlockView world, BlockPos pos, Fluid fluid, Direction direction) {
+		return false;
+	}
+
+	@Override
+	public int getTickRate(WorldView world) {
+		return 5;
+	}
+
+	@Override
+	protected float getBlastResistance() {
+		return 100;
+	}
+}
