+++ fabric-fluid-extension-api-v1/src/testmod/resources/fabric.mod.json
@@ -1,0 +1,16 @@
+{
+  "schemaVersion": 1,
+  "id": "fabric-fluid-extension-api-v1-testmod",
+  "name": "Fabric Fluid Extension API (v1) Test Mod",
+  "version": "1.0.0",
+  "environment": "*",
+  "license": "Apache-2.0",
+  "depends": {
+    "fabric-fluid-extension-api-v1": "*"
+  },
+  "entrypoints": {
+    "main": [
+      "net.fabricmc.fabric.test.fluid.extension.FluidTest"
+    ],
+  }
+}
