--- fabric-fluid-extensions-v1/src/main/resources/fabric.mod.json
@@ -1,26 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "fabric-fluid-extensions-v1",
-  "name": "Fabric Fluid Extensions (v1)",
-  "version": "${version}",
-  "environment": "*",
-  "license": "Apache-2.0",
-  "icon": "assets/fabric-fluid-extensions-v1/icon.png",
-  "contact": {
-    "homepage": "https://fabricmc.net",
-    "irc": "irc://irc.esper.net:6667/fabric",
-    "issues": "https://github.com/FabricMC/fabric/issues",
-    "sources": "https://github.com/FabricMC/fabric"
-  },
-  "authors": [
-    "FabricMC"
-  ],
-  "depends": {
-    "fabricloader": ">=0.9.2",
-    "minecraft": ">=1.16"
-  },
-  "description": "Extensions to assist in creating new fluids.",
-  "mixins": [
-    "fabric-fluid-extensions-v1.mixins.json"
-  ]
-}
