--- fabric-fluid-extensions-v1/src/testmod/java/net/fabricmc/fabric/test/fluid/extension/LemonadeFluid.java
@@ -1,67 +1,0 @@
-package net.fabricmc.fabric.test.fluid.extension;
-
-import net.fabricmc.fabric.api.fluid.extension.v1.FabricFlowableFluid;
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockState;
-import net.minecraft.block.entity.BlockEntity;
-import net.minecraft.fluid.Fluid;
-import net.minecraft.fluid.FluidState;
-import net.minecraft.item.Item;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.Direction;
-import net.minecraft.world.BlockView;
-import net.minecraft.world.WorldAccess;
-import net.minecraft.world.WorldView;
-
-public class LemonadeFluid extends FabricFlowableFluid {
-	@Override
-	public Block getFluidBlock() {
-		return FluidTest.LEMONADE_BLOCK;
-	}
-
-	@Override
-	public int getLevels() {
-		return 20;
-	}
-
-	@Override
-	protected boolean isInfinite() {
-		return false;
-	}
-
-	@Override
-	protected void beforeBreakingBlock(WorldAccess world, BlockPos pos, BlockState state) {
-		final BlockEntity blockEntity = state.getBlock().hasBlockEntity() ? world.getBlockEntity(pos) : null;
-		Block.dropStacks(state, world, pos, blockEntity);
-	}
-
-	@Override
-	protected int getFlowSpeed(WorldView world) {
-		return 4;
-	}
-
-	@Override
-	protected int getLevelDecreasePerBlock(WorldView world) {
-		return 1;
-	}
-
-	@Override
-	public Item getBucketItem() {
-		return FluidTest.LEMONADE_BUCKET;
-	}
-
-	@Override
-	protected boolean canBeReplacedWith(FluidState state, BlockView world, BlockPos pos, Fluid fluid, Direction direction) {
-		return false;
-	}
-
-	@Override
-	public int getTickRate(WorldView world) {
-		return 5;
-	}
-
-	@Override
-	protected float getBlastResistance() {
-		return 100;
-	}
-}
