--- fabric-fluid-extensions-v1/src/testmod/resources/fabric.mod.json
@@ -1,16 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "fabric-fluid-extensions-v1-testmod",
-  "name": "Fabric Fluid Extensions (v1) Test Mod",
-  "version": "1.0.0",
-  "environment": "*",
-  "license": "Apache-2.0",
-  "depends": {
-    "fabric-fluid-extensions-v1": "*"
-  },
-  "entrypoints": {
-    "main": [
-      "net.fabricmc.fabric.test.fluid.extension.FluidTest"
-    ],
-  }
-}
