+++ settings.gradle
@@ -24,7 +24,7 @@ include 'fabric-crash-report-info-v1'
 //include 'fabric-dimensions-v1'
 include 'fabric-events-interaction-v0'
 include 'fabric-events-lifecycle-v0'
-include 'fabric-fluid-extensions-v1'
+include 'fabric-fluid-extension-api-v1'
 include 'fabric-game-rule-api-v1'
 include 'fabric-item-api-v1'
 include 'fabric-item-groups-v0'
